refactor: Done majority of corrections.
Done majority of corrections
