+++ technical-test/src/main/java/com/inditex/technicaltest/application/services/PricesServiceImpl.java
@@ -1,7 +1,7 @@
 package com.inditex.technicaltest.application.services;
 
-import com.inditex.technicaltest.common.enums.InditexExceptionsEnum;
-import com.inditex.technicaltest.common.exceptions.InditexException;
+import com.inditex.technicaltest.domain.common.enums.InditexExceptionsEnum;
+import com.inditex.technicaltest.domain.common.exceptions.InditexException;
 import com.inditex.technicaltest.domain.model.PriceModel;
 import com.inditex.technicaltest.domain.port.in.PricesService;
 import com.inditex.technicaltest.domain.port.out.PricesRepo;
@@ -10,7 +10,7 @@ import lombok.SneakyThrows;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
-import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -20,18 +20,11 @@ public class PricesServiceImpl implements PricesService {
     @Override
     @SneakyThrows
     public PriceModel getByProductIdBrandIdAndApplicationDate(Long productId, Long brandId, LocalDateTime applicationDate) {
-        List<PriceModel> priceModels = pricesRepo.findByProductIdBrandIdAndApplicationDate(productId, brandId, applicationDate);
+        PriceModel priceModel = pricesRepo.findByProductIdBrandIdAndApplicationDate(productId, brandId, applicationDate);
 
-        if(priceModels.isEmpty())
+        if(Objects.isNull(priceModel))
             throw new InditexException(InditexExceptionsEnum.ERROR_701_EMPTY_PRICES.getMessageDto());
 
-        PriceModel finalPriceModel = null;
-
-        for (PriceModel p : priceModels) {
-            if(finalPriceModel == null || finalPriceModel.getPriority() < p.getPriority())
-                finalPriceModel = p;
-        }
-
-        return finalPriceModel;
+        return priceModel;
     }
 }
