--- technical-test/src/main/java/com/inditex/technicaltest/infrastructure/repository/BrandsRepository.java
@@ -1,7 +1,0 @@
-package com.inditex.technicaltest.infrastructure.repository;
-
-import com.inditex.technicaltest.infrastructure.repository.entity.BrandEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BrandsRepository extends JpaRepository<BrandEntity, Long> {
-}
