+++ technical-test/src/main/java/com/inditex/technicaltest/infrastructure/repository/PricesRepository.java
@@ -6,16 +6,17 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Repository
 public interface PricesRepository extends JpaRepository<PriceEntity, Long> {
     @Query(value = """
-            SELECT P.* 
+            SELECT P.*
             FROM PRICES P
             WHERE PRODUCT_ID = ?1
             AND BRAND_ID = ?2
             AND ?3 BETWEEN START_DATE AND END_DATE
+            ORDER BY PRIORITY DESC
+            FETCH FIRST ROW ONLY
             """, nativeQuery = true)
-    List<PriceEntity> findByProductIdBrandIdAndApplicationDate(Long productId, Long brandId, LocalDateTime applicationDate);
+    PriceEntity findByProductIdBrandIdAndApplicationDate(Long productId, Long brandId, LocalDateTime applicationDate);
 }
