--- technical-test/src/main/java/com/inditex/technicaltest/infrastructure/repository/ProductsRepository.java
@@ -1,9 +1,0 @@
-package com.inditex.technicaltest.infrastructure.repository;
-
-import com.inditex.technicaltest.infrastructure.repository.entity.ProductEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProductsRepository extends JpaRepository<ProductEntity, Long> {
-}
