--- technical-test/src/test/java/com/inditex/technicaltest/infrastructure/repository/ProductsRepositoryTest.java
@@ -1,30 +1,0 @@
-package com.inditex.technicaltest.infrastructure.repository;
-
-import com.inditex.technicaltest.infrastructure.repository.entity.ProductEntity;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.context.ActiveProfiles;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-@DataJpaTest
-@ActiveProfiles("test")
-class ProductsRepositoryTest {
-    @Autowired
-    private ProductsRepository productsRepository;
-
-    @Test
-    void Given_HappyPath_When_FindById_Then_Success(){
-        ProductEntity expected = productsRepository.save(getProductEntity());
-
-        ProductEntity actual = productsRepository.findById(expected.getId()).orElse(null);
-
-        assertThat(actual).isNotNull();
-        assertEquals(expected, actual);
-    }
-    private ProductEntity getProductEntity() {
-        return new ProductEntity(1L, "SHOES");
-    }
-}
